Working on builder

+++ DaylinMicroservices-CICD-Builder/src/main/kotlin/daylightnebula/daylinmicroservices/cicd/Builder.kt
@@ -2,11 +2,21 @@ package daylightnebula.daylinmicroservices.cicd
 
 import daylightnebula.daylinmicroservices.Microservice
 import daylightnebula.daylinmicroservices.MicroserviceConfig
+import daylightnebula.daylinmicroservices.filesysteminterface.HashUtils
+import daylightnebula.daylinmicroservices.filesysteminterface.pushFile
+import daylightnebula.daylinmicroservices.filesysteminterface.requestFile
+import daylightnebula.daylinmicroservices.requests.request
+import okhttp3.internal.wait
 import org.json.JSONObject
 import java.io.File
 import java.lang.Exception
 import java.lang.IllegalArgumentException
 import java.lang.Thread.sleep
+import java.security.MessageDigest
+import java.util.*
+import kotlin.collections.HashMap
+import kotlin.concurrent.thread
+import kotlin.random.Random
 
 val endpoints = hashMapOf<String, (json: JSONObject) -> JSONObject>(
     "run_build" to { json ->
@@ -17,6 +27,7 @@ lateinit var service: Microservice
 
 // control variables
 var buildConfig: JSONObject? = null
+var buildUUID = UUID.randomUUID()
 val startTime = System.currentTimeMillis()
 
 fun main(inArgs: Array<String>) {
@@ -27,16 +38,15 @@ fun main(inArgs: Array<String>) {
     val localModeEnabled = args.containsKey("local")
 
     // only create service when not in local mode
-    if (!localModeEnabled) {
-        service = Microservice(
-            MicroserviceConfig(
-                name = "CICD-Builder",
-                tags = listOf("cicd"),
-            ),
-            endpoints
-        )
-        service.start()
-    }
+    service = Microservice(
+        MicroserviceConfig(
+            name = "CICD-Builder",
+            tags = listOf("cicd"),
+        ),
+        if (localModeEnabled) hashMapOf() else endpoints
+    )
+    service.start()
+    sleep(5000)
 
     // if in local mode, load config given path
     if (localModeEnabled) {
@@ -54,26 +64,73 @@ fun main(inArgs: Array<String>) {
 }
 
 // function that runs the build
+var buildUploadDebug = true
 fun runBuild(config: JSONObject) {
     // create and clear build directory
     val buildFolder = File(System.getProperty("user.dir"), "build")
     buildFolder.mkdirs()
-    buildFolder.listFiles()?.forEach { it.deleteRecursively() }
 
-    // clone the repo
-    if (!config.has("url")) throw IllegalArgumentException("URL must be specified for clone")
-    val cloneResult = cloneGitRepo(buildFolder, config.getString("url"), config.optString("auth"))
-    if (cloneResult != 0) throw Exception("Clone operation failed, URL: ${config.getString("url")}, exit code: $cloneResult")
+    if (!buildUploadDebug) {
+        // delete old build
+        buildFolder.listFiles()?.forEach { it.deleteRecursively() }
 
-    // run each command, verifying after each
-    for (any in (config.optJSONArray("commands") ?: throw IllegalArgumentException("Commands to build must be specified"))) {
-        val command = any as? String ?: throw IllegalArgumentException("Commands must be a string")
-        val cmdResult = runCommand(buildFolder, command)
-        if (cmdResult != 0) throw Exception("Command $command could not be run, exited with code $cmdResult")
+        // clone the repo
+        if (!config.has("url")) throw IllegalArgumentException("URL must be specified for clone")
+        val cloneResult = cloneGitRepo(buildFolder, config.getString("url"), config.optString("auth"))
+        if (cloneResult != 0) throw Exception("Clone operation failed, URL: ${config.getString("url")}, exit code: $cloneResult")
+
+        // run each command, verifying after each
+        for (any in (config.optJSONArray("commands")
+            ?: throw IllegalArgumentException("Commands to build must be specified"))) {
+            val command = any as? String ?: throw IllegalArgumentException("Commands must be a string")
+            val cmdResult = runCommand(buildFolder, command)
+            if (cmdResult != 0) throw Exception("Command $command could not be run, exited with code $cmdResult")
+        }
     }
+
+    // for each resource, save it to the file system accordingly
+    config.optJSONArray("resources")?.forEach {
+        val resource = it as? String ?: throw IllegalArgumentException("Resource must be a string")
+        val resourceFile = File(buildFolder, resource)
+        if (!resourceFile.exists()) throw IllegalArgumentException("Resource $resource does not exist, searched at ${resourceFile.absolutePath}")
+        saveResource(resourceFile)
+    } ?: throw IllegalArgumentException("Resources must be specified")
 
     // catch errors from shutdown
     try { service.dispose() } catch (ex: Exception) {}
+}
+
+// function to log and save a resource to the file system
+fun saveResource(resource: File) {
+    // file paths
+    val masterFolderPath = "/.builds/${resource.nameWithoutExtension}"
+    val metadataFilePath = "${masterFolderPath}/metadata.json"
+    val buildPath = "$masterFolderPath/$buildUUID"
+
+    // request metadata file
+    println("Request metadata file $metadataFilePath")
+    service.requestFile(metadataFilePath).get().let { file ->
+        // get json object from file
+        val metadata = file?.let { JSONObject(it.readText()) }
+            ?: JSONObject()
+
+        // update json
+        metadata.put(
+            buildUUID.toString(),
+            JSONObject()
+                .put("id", buildUUID)
+                .put("time", startTime)
+        )
+
+        // create temp file
+        val tempFile = File("tmp-${Random.nextInt(Int.MIN_VALUE, Int.MAX_VALUE)}.json")
+        tempFile.writeText(metadata.toString(4))
+
+        // push
+        println("Push file $metadataFilePath")
+        val pushResult = service.pushFile(metadataFilePath, tempFile).whenComplete { b, _ -> println("Success? $b") }.get()
+        println("Push result $pushResult")
+    }
 }
 
 // function that runs the given command
